refactor : 경매 물건 이미지 저장 클래스 수정 - 53 이미지 등록 기능 메소드 별도 클래스로 분리 - 잘린 cdn 주소 DB에 저장하는 메소드 리턴 값 List로 변경 (#30
